Respect semgrepignore with reachability
